# functions/swagger.yaml
openapi: 3.0.0
info:
  title: RMS API
  version: 1.0.0
  description: Restaurant Management System API
servers:
  - url: https://api-d4o6tbc5fq-uc.a.run.app/
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /reservation/all:
    get:
      summary: Get all reservations
      tags:
        - Reservation
      description: Retrieve all reservations (owner only).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    userId:
                      type: string
                    tableId:
                      type: string
                    startTime:
                      type: string
                    endTime:
                      type: string
                    people:
                      type: integer
                    status:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not an owner
...
  /user/all:
    get:
      summary: Retrieve all users
      tags:
        - User
      description: Retrieve a list of all users. Only accessible by owners.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
                    privileges:
                      type: string
                    phoneNumber:
                      type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /walls:
    get:
      summary: Retrieve a list of all walls
      security:
        - bearerAuth: []
      tags:
        - Walls
      responses:
        '200':
          description: A list of walls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wall'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new wall
      security:
        - bearerAuth: []
      tags:
        - Walls
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Wall'
      responses:
        '201':
          description: Wall created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The ID of the newly created wall
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /walls/{wallId}:
    get:
      summary: Retrieve a wall by ID
      security:
        - bearerAuth: []
      tags:
        - Walls
      parameters:
        - in: path
          name: wallId
          required: true
          schema:
            type: string
          description: The ID of the wall to retrieve
      responses:
        '200':
          description: The wall details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wall'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Wall not found
    put:
      summary: Update a wall by ID
      security:
        - bearerAuth: []
      tags:
        - Walls
      parameters:
        - in: path
          name: wallId
          required: true
          schema:
            type: string
          description: The ID of the wall to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Wall'
      responses:
        '200':
          description: Wall updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wall'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Wall not found
    delete:
      summary: Delete a wall by ID
      security:
        - bearerAuth: []
      tags:
        - Walls
      parameters:
        - in: path
          name: wallId
          required: true
          schema:
            type: string
          description: The ID of the wall to delete
      responses:
        '200':
          description: Wall deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Wall not found
  /windows:
    get:
      summary: Retrieve a list of all windows
      security:
        - bearerAuth: []
      tags:
        - Windows
      responses:
        '200':
          description: A list of windows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Window'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new window
      security:
        - bearerAuth: []
      tags:
        - Windows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Window'
      responses:
        '201':
          description: Window created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The ID of the newly created window
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /windows/{windowId}:
    get:
      summary: Retrieve a window by ID
      security:
        - bearerAuth: []
      tags:
        - Windows
      parameters:
        - in: path
          name: windowId
          required: true
          schema:
            type: string
          description: The ID of the window to retrieve
      responses:
        '200':
          description: The window details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Window'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Window not found
    put:
      summary: Update a window by ID
      security:
        - bearerAuth: []
      tags:
        - Windows
      parameters:
        - in: path
          name: windowId
          required: true
          schema:
            type: string
          description: The ID of the window to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Window'
      responses:
        '200':
          description: Window updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Window'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Window not found
    delete:
      summary: Delete a window by ID
      security:
        - bearerAuth: []
      tags:
        - Windows
      parameters:
        - in: path
          name: windowId
          required: true
          schema:
            type: string
          description: The ID of the window to delete
      responses:
        '200':
          description: Window deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Window not found
...